requirements/base.txt
djangorestframework
djangorestframework-simplejwt
redis
